+++ frontend/src/app/admin/admin-routing.module.ts
@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AdminUsersRolesComponent } from './users-and-roles/admin-users-roles.component';
-import { AdminRoleDetailsComponent } from './users-and-roles/roles/details/admin-role-details.component';
-import { AdminRolesListComponent } from './users-and-roles/roles/list/admin-roles-list.component';
-import { AdminUsersListComponent } from './users-and-roles/users/admin-users.component';
+import { AdminUsersRolesComponent } from './accounts/admin-accounts.component';
+import { AdminRoleDetailsComponent } from './accounts/roles/details/admin-role-details.component';
+import { AdminRolesListComponent } from './accounts/roles/list/admin-roles-list.component';
+import { AdminUsersListComponent } from './accounts/users/admin-users.component';
 import { AdminComponent } from './admin.component';
 
 const routes: Routes = [
   {
     path: '',
+    title: 'Site Administration',
     component: AdminComponent
   },
   {
